requirements.txt
lightning
transformers
lineflow
hydra-core
torchmetrics
numpy
pandas